A big yet interim patch emerging from my attemp...
A big yet interim patch emerging from my attempts to centralize common
classloader-related code. As it turns out, not that much of the
patch is directly associated with that. Most of it is cleanup in the
neighborhoods I was visiting, but there are a few new library files
about which I'm open to feedback:

  scala/util/control/Exception - lots of exception handling code.
  scala/net/Utility - what would be the first file in scala.net.*,
    more code to follow if that sounds like a good package idea.
  scala/util/ScalaClassLoader - mostly convenience methods right
    now, more sophistication to come

Also, this adds a :jar command to the repl which adds a jar to your
classpath and replays the session.