Handled action items (1) and (4) of the phaller...
Handled action items (1) and (4) of the phaller/cunei/rytz agenda
for a less regressive partest, to the extent possible without some
accompanying sbaz changes. Unnecessary sanity check which failed on sane
but non-trunk filesystem layouts removed. Files with custom command
sequences explicitly qualify the path to the filenames listed in the
.cmds file. Review by phaller.