SI-6288 Fix positioning of label jumps
ICode generation was assigning the position of
the last label jump to all jumps to that particular
label def.

This problem is particularly annoying under the new
pattern matcher: a breakpoint in the body of the final
case will be triggered on the way out of the body of
any other case.

Thanks to @dragos for the expert guidance as we
wended our way through GenICode to the troublesome
code. Chalk up another bug for mutability.

I believe that the ICode output should be stable
enough to use a a .check file, if it proves otherwise
we should make it so.