SI-7729 Fix broken windows build
Fixing #ifdiff uncovered a latent bug in partest when it detects
paths to normalize in test log output.  It would not normalize
file separators for test names with non-word characters.

This commit normalizes the test name to the ticket number as a
workaround until the next partest version is picked up.

It also fixes partest in this repo in case we've not yet switched
over to the new repo.  A separate PR will be submitted to the new
repo as necessary.