Upgrade to Scala 2.11.0-M6 and mark Scala dependencies as provided.
As experience has shown, having Scala transitive dependencies from
partest is problematic. The main reason is that we want to test a newer
version of the code (e.g. scala-xml module) that partest depends on.
If partest has transitive dependency we have to go through process of
excluding it's dependency so we are sure that our binaries for given
dependency are picked up. This commit reverses the defaults: all Scala
dependencies are marked as provided which means a consumer of partest
must explicitly provide those dependencies.

A side benefit is that, if one doesn't need scalacheck (e.g. there
are no scalacheck style partest tests) then scalacheck won't show up
on test classpath because it won't be provided.