Make ant a provided dependency
Ant is only needed to compile the Partest ant task. At runtime, the
task is run by the Ant runner, so ant.jar and friends are already
provided.

In the scala build, this removes ant from the compile classpath of
partest-extras, and from the runtime classpath of test.junit and the
`test/partest` script. Nothing in partest other than the task depends
on Ant, so there is no need to explicitly provide ant anywhere. Both
`ant test.run` and `test/partest` should work fine with this change.

The motivation for this is to prevent ant from appearing several time on
the classpath, see [1].

[1]: https://groups.google.com/d/topic/scala-internals/EuMX5xu347o/discussion