Locate partest-javaagent in classpath
It used to be searched in ./build but that doesn’t work for the sbt
build, where it is actually under ./build-sbt. The correct path is
already on the partest classpath, so we can simply pull it from there
with `findArtifact` instead of using `findJar`.

Some other obsolete uses of `findJar` along with `buildPackLibDir` are
also removed.