Mixed Java/Scala compilation like sbt's "Mixed" compileOrder
Partest used to follow an old recommendation for mixed
compilation: first all files are passed to scalac, then
the Java files to javac (with the Scala classfiles on the
classpath), and finally the Scala files again to the Scala
compiler (with the existing classfiles on the classpath).

SBT and the other build tools stop after the second round,
so it makes sense for our partest to do the same by default.

Separate compilation (Java then Scala) can still be tested
using filename groups (`_N`).